+++ src/main/kotlin/io/github/gunpowder/entities/ConfirmPopup.kt
@@ -1,5 +1,6 @@
 package io.github.gunpowder.entities
 
+import io.github.gunpowder.api.components.with
 import io.github.gunpowder.commands.ConfirmCommand
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.text.Text
@@ -7,6 +8,6 @@ import net.minecraft.text.Text
 class ConfirmPopup(val message: Text, val player: ServerPlayerEntity, val callback: () -> Unit) {
     fun show() {
         player.sendMessage(message, false)
-        ConfirmCommand.set(player, callback)
+        player.with<SignPlayerComponent>().confimCallback = callback
     }
 }