+++ src/main/kotlin/io/github/gunpowder/signtypes/SellSign.kt
@@ -28,32 +28,28 @@ import io.github.gunpowder.GunpowderSignshopModule
 import io.github.gunpowder.api.GunpowderMod
 import io.github.gunpowder.api.builders.SignType
 import io.github.gunpowder.api.builders.Text
-import io.github.gunpowder.api.module.currency.modelhandlers.BalanceHandler
+import io.github.gunpowder.api.components.with
 import io.github.gunpowder.entities.ConfirmPopup
+import io.github.gunpowder.entities.SignPlayerComponent
+import io.github.gunpowder.modelhandlers.BalanceHandler
 import net.minecraft.block.entity.*
 import net.minecraft.inventory.Inventories
-import net.minecraft.inventory.Inventory
 import net.minecraft.inventory.SimpleInventory
 import net.minecraft.item.ItemStack
-import net.minecraft.nbt.CompoundTag
-import net.minecraft.nbt.NbtHelper
-import net.minecraft.nbt.NbtOps
+import net.minecraft.nbt.NbtCompound
 import net.minecraft.text.LiteralText
 import net.minecraft.util.Formatting
 import net.minecraft.util.Identifier
-import net.minecraft.util.ItemScatterer
-import net.minecraft.util.collection.DefaultedList
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.math.Direction
 import net.minecraft.util.registry.Registry
 import net.minecraft.util.registry.RegistryKey
 import java.util.*
-import java.util.function.Predicate
 
 object SellSign {
     val dataCache = mutableMapOf<SignBlockEntity, SignSellData>()
     val handler by lazy {
-        GunpowderMod.instance.registry.getModelHandler(BalanceHandler::class.java)
+        BalanceHandler
     }
 
     fun build() {
@@ -115,21 +111,22 @@ object SellSign {
             }
 
             onCreated { signBlockEntity, serverPlayerEntity ->
-                val lastEntity = GunpowderSignshopModule.lastClickCache[serverPlayerEntity.uuid]
+                val comp = serverPlayerEntity.with<SignPlayerComponent>()
+                val lastEntity = comp.selected
                 if (lastEntity != null) {
                     val stack = lastEntity.getStack(0)
                     if (stack.isEmpty) {
                         serverPlayerEntity.sendMessage(LiteralText("No item in the first slot!"), false)
                         signBlockEntity.world?.removeBlock(signBlockEntity.pos, false)
                     } else {
-                        val price = signBlockEntity.text.filter { it.asString().contains("$") }.map { it.asString().replace("$", "").toDoubleOrNull() }
+                        val price = signBlockEntity.texts.filter { it.asString().contains("$") }.map { it.asString().replace("$", "").toDoubleOrNull() }
                         if (price.isEmpty() || price.first() == null || price.first()!! <= 0) {
                             serverPlayerEntity.sendMessage(LiteralText("No price configured on the sign. Make sure to enter '$' on the line with the price."), false)
                             signBlockEntity.world?.removeBlock(signBlockEntity.pos, false)
                         } else {
                             serverPlayerEntity.sendMessage(LiteralText("Put up for sale: $stack for $${price.first()!!}"), false)
                             val data = SignSellData(lazy { lastEntity }, serverPlayerEntity.uuid, stack.copy(), price.first()!!)
-                            GunpowderSignshopModule.lastClickCache.remove(serverPlayerEntity.uuid)
+                            comp.selected = null
                             dataCache[signBlockEntity] = data
                         }
                     }
@@ -147,10 +144,10 @@ object SellSign {
 
             serialize { signBlockEntity, compoundTag ->
                 val data = dataCache[signBlockEntity]  ?: return@serialize
-                val link = CompoundTag()
-                val item = CompoundTag()
+                val link = NbtCompound()
+                val item = NbtCompound()
                 val container by data.linkedContainer
-                data.targetStack.toTag(item)
+                data.targetStack.writeNbt(item)
                 link.put("item", item)
                 link.putUuid("owner", data.ownerUUID)
                 link.putDouble("price", data.price)
@@ -164,9 +161,9 @@ object SellSign {
             deserialize { signBlockEntity, compoundTag ->
                 val link = compoundTag.getCompound("link")
                 val data = SignSellData(
-                    lazy { GunpowderMod.instance.server.getWorld(RegistryKey.of(Registry.DIMENSION, Identifier(link.getString("world"))))?.getBlockEntity(BlockPos(link.getInt("x"), link.getInt("y"), link.getInt("z"))) as ChestBlockEntity },
+                    lazy { GunpowderMod.instance.server.getWorld(RegistryKey.of(Registry.WORLD_KEY, Identifier(link.getString("world"))))?.getBlockEntity(BlockPos(link.getInt("x"), link.getInt("y"), link.getInt("z"))) as ChestBlockEntity },
                     link.getUuid("owner"),
-                    ItemStack.fromTag(link.getCompound("item")),
+                    ItemStack.fromNbt(link.getCompound("item")),
                     link.getDouble("price")
                 )
                 dataCache[signBlockEntity] = data